GPII-1800:  Disentangled node-specific and browser-safe parts of encoding/decoding for wider reuse.
