+++ mainIncludes.js
@@ -1,6 +1,7 @@
 "use strict";
 
 require("./src/js/lib/querystring-coding");
+require("./src/js/lib/querystring-coding-middleware");
 require("./src/js/routable");
 require("./src/js/express");
 require("./src/js/middleware");
